Sudoku grid tests
Implemented tests for the focus arrow logic. This requires making the grid properties public. We can use test cases to assert that the correct cell will be chosen for any arrow key that is not at the edge, and then a separate test for each case of an arrow key at the edge of the grid.