Validate a set of 1-9 cells
This is building toward an implementation of a sudoku puzzle validator. For each row, column and block we need to be able to check that the value 1-9 are represented. To allow this I utilized TDD to verify that a given cell group has the value 1-9 represented, with no duplciates or missing values.