Validate and visually show a completed grid
We may change the UI implementation for this once we have a more completed sudoku puzzle solver, but for now I am simply adding a button that will color the grid if the user has entered a valid sudoku puzzle solution. This allows us to visually confirm that the sudoku puzzle validator is working successfully.