Differentiate between incomplete and invalid grid states
The next step in the challenge is to implement a puzzle solver. To do this we need to be able to understand if each placement contributes to a valid solution or not. This means we need to not only check for a finished grid, but check if the current placements are valid.

We already check separately for duplicate results vs empty results, so this is as simple as updating the methods to return an appropriate validation state instead of a boolean

This required fixing the example incomplete test grids, as I had originally copied them from the invalid grids and added 0s, whereas now they need to be copied from success grids with added 0s instead