Allow for validating individual value placements
When we implement the solver algorithm, we would like to have a more performance way of checking each new solution attempt. This new method allows us to simply check the new value against the row, column and block it exists on instead of checking the whole grid.

This requires a slight modification to the validate cell group logic in which nulls and 0s do not exit the method, so that if the cell group has duplicates it should always return invalid regardless of the state of the other cells. One test case was removed in which a duplicate numbered cell group was expected to be incomplete