Programmatically solve a sodoku puzzle
We want to be able to solve a sudoku puzzle, which is the next step of the challenge. To do this we are using the backtracking algorithm, which involves iterating through until we find a valid solution. We can use the new ValidateCellPlacement method to improve performance by removing the need to iterate over the whole grid for each check

It may be beneficial to utilize more detailed unit tests to verify and describe the behavior, however this would require some method of opening up the internal behaviors of the solver, which I was reluctant to do at first since that would allow other classes to manipulate it's state, but I'm thinking it may be worth adding here

It's not ideal that our unit tests are complex enough that an additional test is needed to verify the assertion logic, but the alternative would be to list out comparisons for the 81 cells of a grid, which would be even more unreadable. In this case we still get to retain our numbered grid layout in the member data that allows for visually confirming the sudoku puzzles as needed